Change type int -> i64 for TClusterResources::{Node,Chunk}Count.

+++ yt/server/security_server/cluster_resources.h
@@ -20,10 +20,10 @@ struct TClusterResources
     TClusterResources();
 
     //! Set node count.
-    TClusterResources&& SetNodeCount(int nodeCount) &&;
+    TClusterResources&& SetNodeCount(i64 nodeCount) &&;
 
     //! Set chunk count.
-    TClusterResources&& SetChunkCount(int chunkCount) &&;
+    TClusterResources&& SetChunkCount(i64 chunkCount) &&;
 
     //! Set tablet count.
     TClusterResources&& SetTabletCount(int tabletCount) &&;
@@ -45,10 +45,10 @@ struct TClusterResources
     /*!
      *  Branched copies are also counted.
      */
-    int NodeCount;
+    i64 NodeCount;
 
     //! Number of chunks created at master.
-    int ChunkCount;
+    i64 ChunkCount;
 
     //! Number of tablets.
     int TabletCount;
@@ -79,8 +79,8 @@ public:
     TClusterResources ToClusterResources(const NChunkServer::TChunkManagerPtr& chunkManager) const;
 
 private:
-    int NodeCount_ = 0;
-    int ChunkCount_ = 0;
+    i64 NodeCount_ = 0;
+    i64 ChunkCount_ = 0;
     int TabletCount_ = 0;
     i64 TabletStaticMemory_ = 0;
     THashMap<TString, i64> DiskSpacePerMedium_;
